# fix: Wildcard permissions algorithm performance