--- a/src/Contracts/Wildcard.php
+++ b/src/Contracts/Wildcard.php
@@ -2,10 +2,13 @@
 
 namespace Spatie\Permission\Contracts;
 
+use Illuminate\Database\Eloquent\Model;
+
 interface Wildcard
 {
-    /**
-     * @param  string|Wildcard  $permission
-     */
-    public function implies($permission): bool;
+    public function __construct(Model $record);
+
+    public function getIndex(): array;
+
+    public function implies(string $permission, string $guardName, array $index): bool;
 }