@danharrin 2023-06-03
Hi @stefanriehl @erikn69, sorry for the @, but I since you both contributed to parts of this wildcard permission system then it's good to loop you in and see what you think about these changes. If you are not interested or no longer using this package, please ignore :)