@stefanriehl 2023-06-05
Hi @danharrin,
Unfortunately, I don't have the time to deal with this more intensively at the moment. But basically this was still an open point, especially when it comes to a lot of permission checks. I will have a closer look when I have more time. Thank you for your work and at first glance it looks promising.
Best,
Stefan