@freekmurze 2023-06-12
@drbyte For me this change seems ok, what do you think?

Are there any more things we'd like to do before tagging a new major version?