@drbyte 2023-06-17
@danharrin I see now that `phpstan` is flagging these concerns ...

```
  Line   src/PermissionRegistrar.php
 ------ -------------------------------------------------------------------------------------
  172    Call to an undefined method Illuminate\Database\Eloquent\Model::getWildcardClass().
 ------ -------------------------------------------------------------------------------------

 ------ --------------------------------------------------------------------------------------
  Line   src/WildcardPermission.php
 ------ --------------------------------------------------------------------------------------
  32     Call to an undefined method Illuminate\Database\Eloquent\Model::getAllPermissions().
```

Aside: while reviewing I noted the use of `Eloquent/Model` in `WildcardPermission` constructor.  But I was paying more attention to the choice of `$record` for variable name, which I eventually satisfied myself that I was okay with. And then I didn't go back and think about this.

I'm undecided whether to flag exceptions for phpstan, or to do something different.  Thoughts?