@danharrin 2023-06-17
@drbyte If I were to design this library from scratch, I would add a required interface to the model to allow for this sort of type safety. But since you don't have that and I doubt you would do such a large breaking change, I don't know if there is really another option apart from flagging these exceptions, unless you use `method_exists()`.