@drbyte 2023-06-17
@danharrin 

1. Agreed on the breaking-change factor. Worth separately exploring pros and cons of forcing an interface for all models capable of permissions handling.

2. By any chance, have you run this new code on Octane/Swoole? I'm wondering if a call to the new `forgetWildcardPermissionIndex()` should be included in `clearPermissionsCollection()`?