@danharrin 2023-06-18
> Worth separately exploring pros and cons of forcing an interface for all models capable of permissions handling.

This is ultimately down to maintainer personal preference, are you ok with enforcing an interface in the upgrade guide?

> By any chance, have you run this new code on Octane/Swoole? I'm wondering if a call to the new forgetWildcardPermissionIndex() should be included in clearPermissionsCollection()?

I have not tested it. However, I do not believe it needs to be cleared after every request, since the index is not user-specific and should still useful across sessions and requests. It may even provide a further performance boost, as the index doesn't need to be regenerated. Let me know if you see any dissonance in this logic.